moved formatDate to utils
