--- source/battle-board/battle-board/app/modules/helpers.ts
@@ -1,10 +1,0 @@
-export default function formatDate(dateStr: Date): string {
-  const date = new Date(dateStr);
-  const formattedDate = date.toLocaleDateString("en-US", {
-    day: "numeric",
-    month: "long",
-    year: "numeric",
-  });
-
-  return formattedDate;
-}
